Enhance app, fix bugs

+++ WeatherApp/Model/WeatherModel.swift
@@ -23,8 +23,8 @@ struct WeatherModel: Codable {
 }
 
 struct Coordinate: Codable {
-    var latitude: Float?
-    var lontitude: Float?
+    var latitude: Double?
+    var lontitude: Double?
 
     private enum CodingKeys: String, CodingKey {
         case latitude = "lat"
@@ -35,9 +35,9 @@ struct Coordinate: Codable {
 struct MainData: Codable {
     var humidity: Float?
     var pressure: Float?
-    var temperature: Float?
-    var tempMax: Float?
-    var tempMin: Float?
+    var temperature: Double?
+    var tempMax: Double?
+    var tempMin: Double?
 
     private enum CodingKeys: String, CodingKey {
         case humidity
